# docs/swagger.yaml
swagger: "2.0"
info:
  title: "go-there API"
  description: "This is the API documentation of the [go-there URL shortener](https://github.com/Fraise/go-there)."
  version: "0.3"
  license:
    name: "MIT license"
    url: "https://github.com/Fraise/go-there/blob/master/LICENSE"
basePath: "/"
...
schemes:
  - "https"
  - "http"
security:
  - ApiKeyAuth: []
  - PasswordAuth: []
securityDefinitions:
  ApiKeyAuth:
    type: "apiKey"
    name: "X-Api-Key"
    in: "header"
  BasicAuth:
    type: "basic"
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
paths:
  /api/users:
    post:
      tags:
        - "users"
      summary: "Create a new user"
      description: ""
      operationId: "createUser"
      parameters:
        - in: "body"
          name: "body"
          description: "User that will be created"
          required: true
          schema:
            $ref: "#/definitions/CreateUser"
      responses:
        "200":
          description: "Returns the generated API key"
          schema:
            $ref: "#/definitions/ApiKey"
        "400":
          description: "Invalid input/User already exists"
          schema:
            $ref: "#/definitions/Error"
    get:
      tags:
        - "users"
      summary: "Get the list of all users"
      description: ""
      operationId: "getAllUsers"
      responses:
        "200":
          description: "Returns the generated API key"
          schema:
            $ref: "#/definitions/PartialUserInfoList"
...
definitions:
  CreateUser:
    type: "object"
    properties:
      create_user:
        type: "string"
        example: "alice"
      create_password:
        type: "string"
        example: "superpassword"
...
  PartialUserInfoList:
    type: "array"
    items:
      $ref: "#/definitions/PartialUserInfo"
  PartialUserInfo:
    type: "object"
    properties:
      username:
        type: "string"
        example: "alice"
      is_admin:
        type: "boolean"
        example: false
  ApiKey:
    type: "object"
    properties:
      api_key:
        type: "string"
        example: "bi44RkM4YWwueFE0d2RvTkF5akpJTzpPSC1rbkdMcm91VlA3N01pZkJ1Y0F3PT0="
  Error:
    type: "object"
    properties:
      error:
        type: "string"
        example: "user already exists"
...